Added some comments/physics improvements
Still having some trouble with physics, but it's getting a little bit
better. Distance/bounce calculations are hard. Oddly, it works slightly
better on detecting collision in the first left path if you ramp up the
velocity. Also, commented a bit on file I/O and the collision detection,
since that is also on the rubric.